pyproject.toml
[tool.poetry]
name = "darkman"
version = "0.1.0"
description = ""
authors = ["Hugo Osvaldo Barrera <hugo@barrera.io>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.9"
astral = "^2.2"
python-dateutil = "^2.8.1"
pyxdg = "^0.27"
jeepney = "^0.6.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
